Renamve footer
