+++ src/presentation/web/src/components/navbar/Navbar.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { AppBar, Grid, Toolbar } from '@material-ui/core';
 import NavbarList from '../navbarList/NavbarList';
-import NavbarLoginBtn from '../navbarLoginBtn/NavbarLoginBtn';
+import NavbarLoginBtn from '../navbarLoginButton/NavbarLoginBtn';
 import { Link as RouterLink } from 'react-router-dom';
 import { Link } from '@material-ui/core';
